And more changes to HC. Some more base changes are coming up.

+++ java/k4unl/minecraft/Hydraulicraft/client/renderers/transportation/RendererPartHose.java
@@ -64,7 +64,7 @@ public class RendererPartHose extends TileEntitySpecialRenderer {
 		TileMultipart mp = (TileMultipart)tileentity;
 		if(Multipart.hasPartHose(mp)){
 			PartHose tp = Multipart.getHose(mp);
-			doRender(x, y, z, f, tp.getTier(), tp.getConnectedSides());
+			doRender(x, y, z, f, tp.getTier().toInt(), tp.getConnectedSides());
 		}
 	}
